Finish AlgorithmKNN class
 - Main class has not been implemented

[ticket: patrick_pu]

+++ Algorithm.java
@@ -3,7 +3,7 @@ import java.util.*;
 public interface Algorithm {
 	public void loadDataCenter(DataCenter dc);
 	
-	public int getRatingByUserAndMovie(User user, Movie movie);
+	public double getRatingByUserAndMovie(User user, Movie movie);
 	
 	public List<Movie> getTopNRatingMovies(User user, int n);
 }
